Try to avoid losing the refresh_token
The server rotates the refresh_token each time a new access_token is requested. If for some reason, the response did not make it to the client, the server would invalidate the current refresh_token without the client getting the new one.

To mitigate that, there's a 10min grace period during which the old token is still valid. During these 10min, we now retry to get a valid token several times.